chore(component): abstracted Table.tsx further
