+++ biproxi/src/components/Table/Table.tsx
@@ -1,16 +1,7 @@
 import {TableProps, Todo} from "../../index";
-import {ToDoItem} from "../ToDoItem/ToDoItem";
 
-const parseData = (todoList: Todo[]) => {
-    if (todoList.length === 0) {
-        return <p>No todos</p>
-    }
-    return todoList.map((todo: Todo) => {
-        return <ToDoItem key={todo.createdAt} todo={todo}/>
-    })
-};
 const Table = (props: TableProps) => {
-    const {headers, todos} = props;
+    const {headers, data} = props;
   return (
     <table>
         <thead>
@@ -21,7 +12,7 @@ const Table = (props: TableProps) => {
         </tr>
         </thead>
         <tbody>
-        {parseData(todos)}
+            {data}
         </tbody>
       </table>
   )
